+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix `UniqueIndexChecker` to skip partial indexes (indexes with a `WHERE` clause).
 - Fix `MissingUniqueIndexChecker` to skip uniqueness validators with a `conditions` option.
 - Fix `MissingDependentDestroyChecker` to skip polymorphic associations. Thanks [Andy Allan](https://github.com/gravitystorm) for reporting this!
+- Fix `ForeignKeyCascadeChecker` to handle missing primary key (no-primary-key tables) and support composite foreign/primary keys. Thanks [Andy Allan](https://github.com/gravitystorm) for reporting this!
 
 ### [2.1.3] - 2026/02/20